+++ src/app/components/anmeldungen/anmeldungen.component.html
@@ -1,9 +1,4 @@
 <div class="container d-flex flex-column mt-5">
-  <div class="mb-3 d-flex justify-content-end">
-    <button class="btn btn-outline-primary" (click)="toggleView()">
-      {{ view() === 'participants' ? 'Stammeszuteilung' : 'Teilnehmerliste' }}
-    </button>
-  </div>
   <div [formGroup]="formGroup" class="input-group mb-3">
     @if ($jahre(); as jahre) {
     <select formControlName="selectedYear" class="form-select" [class.text-secondary]="!formGroup.value.selectedYear">
@@ -24,7 +19,6 @@
     </select>
     }
   </div>
-  @if (view() === 'participants') {
   @if($anmeldungen(); as anmeldungen) {
   <ngx-datatable #datatable [rows]="anmeldungen" [messages]="{ emptyMessage: 'Keine Daten', totalMessage: 'insgesamt' }"
     [headerHeight]="50" [rowHeight]="50" [footerHeight]="50" columnMode="force" class="bootstrap">
@@ -143,163 +137,4 @@
   </button>
   }
   }
-  } @else {
-  <div class="d-flex flex-column" style="height: calc(100vh - 150px);">
-
-    <div class="row flex-grow-1 overflow-hidden mt-4">
-
-      <div class="col-lg-2 col-md-3 d-flex flex-column h-100">
-        <div class="card shadow-sm flex-grow-1 overflow-hidden">
-          <div
-            class="card-header bg-primary text-white d-flex justify-content-between align-items-center flex-shrink-0">
-            Alle Teilnehmer
-            <span class="badge bg-white text-primary">{{ filteredParticipants().length }}</span>
-          </div>
-
-          <div class="p-2 bg-light border-bottom flex-shrink-0">
-            <div class="input-group input-group-sm">
-              <span class="input-group-text bg-white border-end-0">
-                <i class="bi bi-search"></i>
-              </span>
-              <input type="text" class="form-control border-start-0 ps-0" placeholder="Suchen..."
-                (input)="updateSearch($event)">
-            </div>
-          </div>
-
-          <div class="card-body p-2 overflow-y-auto" sortable data-list-id="main" (onDrop)="onDrop($event)"
-            style="background: #fafafa;">
-            @for (p of filteredParticipants(); track p.id) {
-            <div class="p-2 mb-1 border rounded shadow-sm cursor-move overflow-hidden {{ getGenderClass(p) }}"
-              [attr.data-id]="p.id" style="font-size: 0.85rem;">
-              <strong class="d-block text-truncate">
-                {{ p.person.domainAttributes.firstName }}
-                {{ p.person.domainAttributes.lastName }}
-              </strong>
-              @if (showIds()) {
-              <small class="text-muted">#{{ p.id }}</small>
-              }
-              @if (showAge()) {
-              <div class="small text-muted">
-                Alter: {{ getAgeThisYear(p.personFields?.birthday) }}
-              </div>
-              }
-            </div>
-            } @empty {
-            <div class="text-center text-muted mt-4 small">
-              Keine Teilnehmer gefunden.
-            </div>
-            }
-          </div>
-        </div>
-      </div>
-
-      <div class="col-md-9 h-100 overflow-y-auto pb-3">
-        <div class="row g-3">
-          @for (group of groups(); track $index) {
-          <div class="col-lg-3 col-md-4 col-sm-6">
-            <div class="card border-primary">
-              <div class="card-header py-1 bg-light d-flex justify-content-between align-items-center">
-                <div>
-                  <span>Stamm {{ $index + 1 }}</span>
-                  <small class="text-muted ms-2">
-                    (Ø {{ getAverageAge(group) }} | Var {{ getAgeVariance(group) }})
-                  </small>
-                </div>
-
-                @let counts = getGenderCounts(group);
-                <small class="fw-bold">
-                  <span class="text-primary">{{ counts.boys }}</span>
-                  <span class="text-muted"> / </span>
-                  <span class="text-danger">{{ counts.girls }}</span>
-                </small>
-              </div>
-
-              <div class="card-body p-2" sortable [attr.data-list-id]="$index" (onDrop)="onDrop($event)"
-                style="min-height: 120px; background: #fff;">
-                @for (item of group; track item.id) {
-                <div class="p-2 mb-1 border rounded shadow-sm cursor-move {{ getGenderClass(item) }}"
-                  [attr.data-id]="item.id" style="font-size: 0.85rem;">
-                  <strong>
-                    {{ item.person.domainAttributes.firstName }}
-                    {{ item.person.domainAttributes.lastName }}
-                  </strong>
-                  @if (showIds()) {
-                  <small class="text-muted">##{{ item.id }}</small>
-                  }
-                  @if (showAge()) {
-                  <div class="small text-muted">
-                    Alter: {{ getAgeThisYear(item.personFields?.birthday) }}
-                  </div>
-                  }
-                </div>
-                }
-              </div>
-            </div>
-          </div>
-          }
-        </div>
-      </div>
-    </div>
-
-    <div class="mt-auto pt-3 border-top d-flex justify-content-between align-items-center bg-white flex-shrink-0">
-      <div class="d-flex align-items-center">
-        <div class="form-check form-switch px-4">
-          <input class="form-check-input" type="checkbox" role="switch" id="idToggle" [checked]="showIds()"
-            (change)="toggleIds()">
-          <label class="form-check-label fw-bold text-muted" for="idToggle">
-            <i class="bi bi-bug me-1"></i> Debug Info
-          </label>
-        </div>
-        <div class="form-check form-switch px-4">
-          <input class="form-check-input" type="checkbox" role="switch" id="ageToggle" [checked]="showAge()"
-            (change)="toggleAge()">
-          <label class="form-check-label fw-bold text-muted" for="ageToggle">
-            <i class="bi bi-bug me-1"></i> Show Age
-          </label>
-        </div>
-      </div>
-
-      <div class="d-flex gap-4 pb-2">
-        <div class="btn-group shadow-sm" role="group">
-          <button class="btn btn-outline-secondary" (click)="loadGroups()">
-            <i class="bi bi-arrow-clockwise me-1"></i> Entwurf laden
-          </button>
-          <button class="btn btn-outline-secondary" (click)="saveGroups()">
-            <i class="bi bi-save me-1"></i> Zwischenspeichern
-          </button>
-        </div>
-
-        <div class="btn-group shadow-sm" role="group">
-          <button class="btn btn-info text-white" (click)="loadGroupsServer()">
-            <i class="bi bi-cloud-download me-1"></i> Server laden
-          </button>
-          @if (showIds()) {
-          <button class="btn btn-warning" (click)="saveGroupsServerSlow()">
-            <i class="bi bi-hourglass-split me-1"></i> Slow Save
-          </button>
-          }
-          <button class="btn btn-success" (click)="saveGroupsServer()">
-            <i class="bi bi-cloud-upload me-1"></i> Final speichern
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-  }
 </div>
-@if ($progress() > 0) {
-<div class="position-fixed top-0 start-0 w-100 h-100 d-flex justify-content-center align-items-center"
-  style="background: rgba(0,0,0,0.5); z-index: 9999;">
-  <div class="card p-4 shadow-lg text-center" style="min-width: 300px;">
-    <div class="spinner-border text-primary mb-3" role="status"></div>
-    <h5 class="mb-2">Speichere Daten...</h5>
-    <p class="text-muted small">Bitte das Fenster nicht schließen.</p>
-
-    <div class="progress" style="height: 10px;">
-      <div class="progress-bar progress-bar-striped progress-bar-animated" role="progressbar"
-        [style.width.%]="$progress() * 100">
-      </div>
-    </div>
-  </div>
-</div>
-}
