+++ src/app/components/anmeldungen/anmeldungen.component.ts
@@ -41,23 +41,6 @@ export class AnmeldungenComponent {
     selectedWeek: this.fb.control<number | null>(null),
   });
 
-  view = signal<'participants' | 'groups'>('participants');
-  groups = signal<AnmeldungenViewModel[][]>(Array.from({ length: 8 }, () => []));
-  toggleView() {
-    this.view.update(v => v === 'participants' ? 'groups' : 'participants');
-  }
-
-
-  showIds = signal<boolean>(false);
-  toggleIds() {
-    this.showIds.update(v => !v);
-  }
-
-  showAge = signal<boolean>(false);
-  toggleAge() {
-    this.showAge.update(v => !v);
-  }
-
   readonly $groupTypes = toSignal(this.churchToolsService.getGroupTypes());
   readonly $jahre = toSignal(this.churchToolsService.getJahre());
 
@@ -68,7 +51,6 @@ export class AnmeldungenComponent {
     debounceTime(1000),
     tap(() => {
       this.formGroup.controls.selectedWeek.reset();
-      this.groups.set(Array.from({ length: 8 }, () => []));
     }),
     filter((value): value is number => !!value),
     switchMap(groupId => this.churchToolsService.getSolawochen(groupId)),
@@ -78,7 +60,6 @@ export class AnmeldungenComponent {
     distinctUntilChanged(),
     debounceTime(1000),
     tap(() => {
-      this.groups.set(Array.from({ length: 8 }, () => []));
     }),
     filter((value): value is number => !!value),
     switchMap(groupId => this.churchToolsService.getAnmeldungen(groupId)),
@@ -195,311 +176,5 @@ export class AnmeldungenComponent {
     } finally {
       setTimeout(() => this.$progress.set(0), 500);
     }
-  }
-
-  sortableOptions = {
-    group: 'nested',
-    animation: 150,
-    fallbackOnBody: true,
-    swapThreshold: 0.65
-  };
-
-  moveParticipant(movedItem: AnmeldungenViewModel, targetList: 'main' | number) {
-    // Remove from everywhere first
-    this.$anmeldungen.update(list => list.filter(p => p.id !== movedItem.id));
-    this.groups.update(groups => groups.map(g => g.filter(p => p.id !== movedItem.id)));
-
-    // Add to target
-    if (targetList === 'main') {
-      this.$anmeldungen.update(list => [...list, movedItem]);
-    } else {
-      this.groups.update(groups => {
-        const newGroups = [...groups];
-        newGroups[targetList] = [...newGroups[targetList], movedItem];
-        return newGroups;
-      });
-    }
-  }
-
-  getGenderClass(p: AnmeldungenViewModel): string {
-    const sexId = p.personFields?.sexId;
-    if (sexId === 1) return 'border-primary border-3 bg-blue-light'; // Boy (Blue)
-    if (sexId === 2) return 'border-danger border-3 bg-red-light';   // Girl (Red)
-    return 'bg-white';
-  }
-
-  private get storageKey(): string {
-    return `groups_assignment_${this.formGroup.value.selectedWeek}`;
-  }
-
-  saveGroups() {
-    const weekId = this.formGroup.value.selectedWeek;
-    if (!weekId) return;
-
-    // We only store the IDs of the persons in each group to keep the storage clean
-    const assignment = this.groups().map(group => group.map(p => p.id));
-
-    localStorage.setItem(`groups_week_${weekId}`, JSON.stringify(assignment));
-    alert('Gruppen lokal gespeichert!');
-  }
-
-  async saveGroupsServer() {
-    const groupId = this.formGroup.value.selectedWeek;
-    if (!groupId || this.$progress() > 0) return;
-    try {
-      const allFields = await firstValueFrom(this.churchToolsService.getGroupMemberFields(groupId));
-      const targetField = allFields.find(f => f.name === 'Stammeszugehörigkeit' || f.name === 'Gruppenzugehörigkeit');
-      if (!targetField) {
-        alert("Zielfeld nicht gefunden!");
-        return;
-      }
-      const fieldId = targetField.id;
-      const fieldName = targetField.name;
-      this.$progress.set(0.01);
-      const tasks: (() => Promise<void>)[] = [];
-      this.groups().forEach((group, i) => {
-        const groupName = `Stamm ${i + 1}`;
-        group.forEach(member => {
-          const currentServerVal = member.fields.find(f => f.id === fieldId || f.name === fieldName)?.value;
-          if (currentServerVal !== groupName) {
-            tasks.push(async () => {
-              const updated = await firstValueFrom(this.churchToolsService.updateGroupMemberFields(groupId, member.personId, { [fieldId]: groupName }));
-              this.groups.update(gs => {
-                const newGs = [...gs];
-                newGs[i] = newGs[i].map(m => m.id === member.id ? { ...m, fields: updated.fields } : m);
-                return newGs;
-              });
-            });
-          }
-        });
-      });
-      this.$anmeldungen().forEach(member => {
-        const currentServerVal = member.fields.find(f => f.id === fieldId || f.name === fieldName)?.value;
-        if (currentServerVal !== null && currentServerVal !== '') {
-          tasks.push(async () => {
-            const updated = await firstValueFrom(this.churchToolsService.updateGroupMemberFields(groupId, member.personId, { [fieldId]: null }));
-            this.$anmeldungen.update(list => list.map(m => m.id === member.id ? { ...m, fields: updated.fields } : m));
-          });
-        }
-      });
-      if (tasks.length === 0) {
-        alert('Alles bereits auf dem neuesten Stand!');
-        this.$progress.set(0);
-        return;
-      }
-      const CHUNK_SIZE = 15;
-      for (let i = 0; i < tasks.length; i += CHUNK_SIZE) {
-        const chunk = tasks.slice(i, i + CHUNK_SIZE);
-        await Promise.all(chunk.map(task => task()));
-        this.$progress.set((i + chunk.length) / tasks.length);
-        await new Promise(resolve => setTimeout(resolve, 100));
-      }
-      alert(`${tasks.length} Änderungen erfolgreich gespeichert!`);
-    } catch (err) {
-      console.error("Speicherfehler", err);
-      alert("Ein Fehler ist aufgetreten beim Speichern.");
-    } finally {
-      this.$progress.set(0);
-    }
-  }
-  async saveGroupsServerSlow() {
-    const groupId = this.formGroup.value.selectedWeek;
-    if (!groupId || this.$progress() > 0) return;
-
-    try {
-      const allFields = await firstValueFrom(this.churchToolsService.getGroupMemberFields(groupId));
-      const targetField = allFields.find(f => f.name === 'Stammeszugehörigkeit' || f.name === 'Gruppenzugehörigkeit');
-      if (!targetField) return;
-      const fieldId = targetField.id;
-      const fieldName = targetField.name;
-      const currentGroups = this.groups();
-      const unassigned = this.$anmeldungen();
-      const updates: { member: AnmeldungenViewModel, newValue: string | null, targetGroupIndex?: number }[] = [];
-      currentGroups.forEach((group, i) => {
-        const groupName = `Stamm ${i + 1}`;
-        group.forEach(member => {
-          const currentServerVal = member.fields.find(f => f.id === fieldId || f.name === fieldName)?.value;
-          if (currentServerVal !== groupName) {
-            updates.push({ member, newValue: groupName, targetGroupIndex: i });
-          }
-        });
-      });
-      unassigned.forEach(member => {
-        const currentServerVal = member.fields.find(f => f.id === fieldId || f.name === fieldName)?.value;
-        if (currentServerVal !== null && currentServerVal !== '') {
-          updates.push({ member, newValue: null });
-        }
-      });
-      if (updates.length === 0) {
-        alert('Keine Änderungen zum Speichern gefunden.');
-        return;
-      }
-      this.$progress.set(0.01);
-      let processed = 0;
-      for (const task of updates) {
-        const updated = await firstValueFrom(
-          this.churchToolsService.updateGroupMemberFields(groupId, task.member.personId, { [fieldId]: task.newValue })
-        );
-        if (task.targetGroupIndex !== undefined) {
-          this.groups.update(gs => {
-            const newGs = [...gs];
-            newGs[task.targetGroupIndex!] = newGs[task.targetGroupIndex!].map(m =>
-              m.id === task.member.id ? { ...m, fields: updated.fields } : m
-            );
-            return newGs;
-          });
-        } else {
-          this.$anmeldungen.update(list => list.map(m =>
-            m.id === task.member.id ? { ...m, fields: updated.fields } : m
-          ));
-        }
-        processed++;
-        this.$progress.set(processed / updates.length);
-        await new Promise(resolve => setTimeout(resolve, 500));
-      }
-      alert(`${updates.length} Änderungen im langsamen Modus abgeschlossen!`);
-    } catch (err) {
-      console.error("Fehler im langsamen Modus", err);
-      alert("Ein Fehler ist im langsamen Modus aufgetreten.");
-    } finally {
-      this.$progress.set(0);
-    }
-  }
-
-  loadGroups() {
-    const weekId = this.formGroup.value.selectedWeek;
-    if (!weekId) return;
-    const savedData = localStorage.getItem(`groups_week_${weekId}`);
-    if (!savedData) {
-      alert('Keine gespeicherten Gruppen gefunden.');
-      return;
-    }
-    const allParticipants = [...this.$anmeldungen(), ...this.groups().flat()];
-    const savedGroupIds: number[][] = JSON.parse(savedData);
-    const flatSavedIds = savedGroupIds.flat();
-    const newGroups: AnmeldungenViewModel[][] = savedGroupIds.map(ids =>
-      allParticipants.filter(p => ids.includes(p.id))
-    );
-    const unassigned = allParticipants.filter(p => !flatSavedIds.includes(p.id));
-    this.groups.set(newGroups);
-    this.$anmeldungen.set(unassigned);
-  }
-
-  loadGroupsServer() {
-    const allParticipants = [...this.$anmeldungen(), ...this.groups().flat()];
-    const newGroups = Array.from({ length: 8 }, () => [] as AnmeldungenViewModel[]);
-    const remainingInMain: AnmeldungenViewModel[] = [];
-    allParticipants.forEach(p => {
-      const groupField = p.fields.find(f =>
-        f.name === 'Stammeszugehörigkeit' || f.name === 'Gruppenzugehörigkeit'
-      );
-      const val = groupField?.value;
-      if (val && typeof val === 'string') {
-        const match = val.match(/\d+/);
-        const groupNum = match ? parseInt(match[0], 10) : null;
-        if (groupNum !== null && groupNum >= 1 && groupNum <= 8) {
-          newGroups[groupNum - 1].push(p);
-        } else {
-          remainingInMain.push(p);
-        }
-      } else {
-        remainingInMain.push(p);
-      }
-    });
-    this.groups.set(newGroups);
-    this.$anmeldungen.set(remainingInMain);
-  }
-
-  onDrop(event: { item: any, from: string, to: string, oldIndex: number, newIndex: number }) {
-    const personId = Number(event.item.getAttribute('data-id'));
-    let movedItem: AnmeldungenViewModel | undefined;
-    if (event.from === 'main') {
-      movedItem = this.$anmeldungen().find(p => p.id === personId);
-      this.$anmeldungen.update(list => list.filter(p => p.id !== personId));
-    } else {
-      const fromIdx = Number(event.from);
-      movedItem = this.groups()[fromIdx].find(p => p.id === personId);
-      this.groups.update(gs => {
-        const newGs = [...gs];
-        newGs[fromIdx] = newGs[fromIdx].filter(p => p.id !== personId);
-        return newGs;
-      });
-    }
-    if (!movedItem) return;
-    if (event.to === 'main') {
-      this.$anmeldungen.update(list => {
-        const newList = [...list];
-        newList.splice(event.newIndex, 0, movedItem!);
-        return newList;
-      });
-    } else {
-      const toIdx = Number(event.to);
-      this.groups.update(gs => {
-        const newGs = [...gs];
-        const targetGroup = [...newGs[toIdx]];
-        targetGroup.splice(event.newIndex, 0, movedItem!);
-        newGs[toIdx] = targetGroup;
-        return newGs;
-      });
-    }
-  }
-
-
-  getGenderCounts(group: AnmeldungenViewModel[]) {
-    return {
-      boys: group.filter(p => p.personFields?.sexId === 1).length,
-      girls: group.filter(p => p.personFields?.sexId === 2).length
-    };
-  }
-
-  getAgeThisYear(birthday: any): number | null {
-    if (!birthday) return null;
-    const birthDate = parseISO(String(birthday));
-    if (!isValid(birthDate)) return null;
-
-    const currentYear = new Date().getFullYear();
-    return currentYear - birthDate.getFullYear();
-  }
-
-  getAverageAge(group: AnmeldungenViewModel[]): number {
-    if (group.length === 0) return 0;
-    const totalAge = group.reduce((sum, member) => {
-      const age = this.getAgeThisYear(member.personFields?.birthday);
-      return sum + (age || 0);
-    }, 0);
-    return Math.round(totalAge / group.length);
-  }
-
-
-  searchTerm = signal<string>('');
-
-  filteredParticipants = computed(() => {
-    const query = this.searchTerm().toLowerCase();
-    const all = this.$anmeldungen(); // This is your original list signal
-
-    if (!query) return all;
-
-    return all.filter(p =>
-      p.person.domainAttributes.firstName.toLowerCase().includes(query) ||
-      p.person.domainAttributes.lastName.toLowerCase().includes(query) ||
-      p.id.toString().includes(query)
-    );
-  });
-
-  updateSearch(event: Event) {
-    const element = event.target as HTMLInputElement;
-    this.searchTerm.set(element.value);
-  }
-
-  getAgeVariance(group: AnmeldungenViewModel[]): number {
-    if (group.length <= 1) return 0;
-    const ages = group
-      .map(m => this.getAgeThisYear(m.personFields?.birthday))
-      .filter((age): age is number => age !== null);
-    if (ages.length === 0) return 0;
-    const mean = ages.reduce((a, b) => a + b, 0) / ages.length;
-    const squaredDiffs = ages.map(age => Math.pow(age - mean, 2));
-    const variance = squaredDiffs.reduce((a, b) => a + b, 0) / ages.length;
-    return Math.round(variance * 10) / 10;
   }
 }
