+++ src/app/components/stammes-management/stammes-management.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { StammesManagementComponent } from './stammes-management.component';
+
+describe('StammesManagementComponent', () => {
+  let component: StammesManagementComponent;
+  let fixture: ComponentFixture<StammesManagementComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [StammesManagementComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(StammesManagementComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
